backend/requirements.txt
mem0ai 
openai 
fastapi 
uvicorn 
python-dotenv
